Create new 'lintPostcssResult' module
In a future PR this module will be required by the new `browser.js` module. This change has been extracted from: #4796
